add private message validation and simple cache

+++ src/AhRulesBot/AhRulesBot/Infrastructure/AppConfig.cs
@@ -5,7 +5,7 @@ namespace AhRulesBot.Infrastructure
     {
         public long TestChatId { get; set; }
         public long AHChatId { get; set; }
-        public long BotAdminId { get; set; }
+        public int BotAdminId { get; set; }
         public string ApiKey { get; set; }
         public string BotName { get; set; }
     }
